+++ resources/views/order.blade.php
@@ -172,17 +172,20 @@ function fillOrder() {
 }
 
 function delOrder() {
-    $.ajax({
-        url: "{{ url('orders', $order->id) }}",
-        type: 'DELETE',
-        data: {
-            _token: '{{csrf_token()}}',
-        },
-        success: function(result) {
-            alert('Delete success!');
-            location.href = '/orders';
-        }
-    });
+    let text = "Are you sure you want to delete this item?";
+    if (confirm(text)) {
+        $.ajax({
+            url: "{{ url('orders', $order->id) }}",
+            type: 'DELETE',
+            data: {
+                _token: '{{csrf_token()}}',
+            },
+            success: function(result) {
+                alert('Delete success!');
+                location.href = '/orders';
+            }
+        });
+    }
 }
 @endif
