+++ resources/views/orders.blade.php
@@ -57,17 +57,20 @@ function completed() {
 }
 
 function del(url) {
-    $.ajax({
-        url: url,
-        type: 'DELETE',
-        data: {
-            _token: '{{csrf_token()}}',
-        },
-        success: function(result) {
-            alert('Delete success!');
-            location.href = '/orders';
-        }
-    });
+    let text = "Are you sure you want to delete this item?";
+    if (confirm(text)) {
+        $.ajax({
+            url: url,
+            type: 'DELETE',
+            data: {
+                _token: '{{csrf_token()}}',
+            },
+            success: function(result) {
+                alert('Delete success!');
+                location.href = '/orders';
+            }
+        });
+    }
 }
 </script>
 @stop
