+++ resources/views/users.blade.php
@@ -64,17 +64,20 @@ function view(url) {
 }
 
 function del(url) {
-    $.ajax({
-        url: url,
-        type: 'DELETE',
-        data: {
-            _token: '{{csrf_token()}}',
-        },
-        success: function(result) {
-            console.log('Delete response: ' + result);
-            location.href = '/users';
-        }
-    });
+    let text = "Are you sure you want to delete this item?";
+    if (confirm(text)) {
+        $.ajax({
+            url: url,
+            type: 'DELETE',
+            data: {
+                _token: '{{csrf_token()}}',
+            },
+            success: function(result) {
+                console.log('Delete response: ' + result);
+                location.href = '/users';
+            }
+        });
+    }
 }
 </script>
 @stop
